+++ src/library/winehook.h
@@ -1,0 +1,58 @@
+/*
+    Copyright 2015-2020 Clément Gallet <clement.gallet@ens-lyon.org>
+
+    This file is part of libTAS.
+
+    libTAS is free software: you can redistribute it and/or modify
+    it under the terms of the GNU General Public License as published by
+    the Free Software Foundation, either version 3 of the License, or
+    (at your option) any later version.
+
+    libTAS is distributed in the hope that it will be useful,
+    but WITHOUT ANY WARRANTY; without even the implied warranty of
+    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+    GNU General Public License for more details.
+
+    You should have received a copy of the GNU General Public License
+    along with libTAS.  If not, see <http://www.gnu.org/licenses/>.
+ */
+
+#ifndef LIBTAS_WINEHOOK_H_INCLUDED
+#define LIBTAS_WINEHOOK_H_INCLUDED
+
+#ifndef __stdcall
+# ifdef __i386__
+#  ifdef __GNUC__
+#   if (__GNUC__ > 4) || ((__GNUC__ == 4) && (__GNUC_MINOR__ >= 2)) || defined(__APPLE__)
+#    define __stdcall __attribute__((__stdcall__)) __attribute__((__force_align_arg_pointer__))
+#   else
+#    define __stdcall __attribute__((__stdcall__))
+#   endif
+#  endif
+# elif defined(__x86_64__) && defined (__GNUC__)
+#  if (__GNUC__ > 5) || ((__GNUC__ == 5) && (__GNUC_MINOR__ >= 3))
+#   define __stdcall __attribute__((ms_abi)) __attribute__((__force_align_arg_pointer__))
+#  else
+#   define __stdcall __attribute__((ms_abi))
+#  endif
+# endif  /* __i386__ */
+#endif /* __stdcall */
+
+struct winstring {
+  unsigned short Length;
+  unsigned short MaximumLength;
+  char *Buffer;
+};
+
+#include "global.h"
+
+namespace libtas {
+
+void hook_ntdll();
+
+OVERRIDE long __stdcall LdrGetProcedureAddress(void *module, const winstring *name,
+                                            unsigned long ord, void **address);
+
+}
+
+#endif